requirements.txt
streamlit
streamlit-audiorecorder
pydub
git+https://github.com/openai/whisper.git 